requirements.txt
# MARS — Multi-Agent Research System
# Requirements (Python 3.11+)
# Last Updated: March 2026

# ── Core agent framework ─────────────────────────────────────────────────────
langchain>=0.3.0,<0.4.0
langgraph>=0.2.0,<0.3.0
langchain-community>=0.3.0,<0.4.0

# ── LLM provider (Groq — free tier, OpenAI-compatible) ──────────────────────
groq>=0.11.0,<0.13.0
langchain-groq>=0.2.0,<0.3.0

# ── Web search ────────────────────────────────────────────────────────────────
tavily-python>=0.5.0,<0.6.0

# ── Web scraping ──────────────────────────────────────────────────────────────
httpx>=0.27.0,<0.29.0
beautifulsoup4>=4.12.0,<5.0.0
playwright>=1.45.0,<2.0.0      # optional: JS-heavy page fallback
# lxml>=5.0.0,<6.0.0            # needs C++ Build Tools — use html.parser instead

# ── UI ────────────────────────────────────────────────────────────────────────
streamlit>=1.35.0,<2.0.0

# ── Observability ──────────────────────────────────────────────────────────────
langsmith>=0.2.0,<0.3.0

# ── Vector DB (optional v2 — requires Rust, skip for now) ─────────────────────
# chromadb>=0.5.0,<0.6.0   # Uncomment in v2 after: pip install chromadb

# ── Utilities ──────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0,<2.0.0
pydantic>=2.0.0,<3.0.0
